requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.4.22
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==6.0.4
django-celery-beat==2.9.0
django-redis==6.0.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-yasg==1.21.15
email-validator==2.3.0
fastapi==0.136.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
inflection==0.5.1
Jinja2==3.1.6
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.2
phonenumbers==9.0.29
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.58.0
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
starlette==1.0.0
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.46.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
